Moved DesertWellDecorator class in the overworld decorators package
